+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * **core**: Create and list classifier jobs (`/classifier-jobs` endpoint)
 * **core**: Access to classifiers can be public (open to all) or private (only the creator)
 * **core**: Dequeue classifier jobs (`POST /internal/classifier-jobs/dequeue`)
+* **core**: Update dequeue classifier jobs to grab jobs age more than or equal to 2 minutes
 
 ### Bug Fixes
 * **noncore**: Fix sensor values table not support Infineon sensor